@@ -1,6 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+#nullable disable
+
+using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
@@ -10,7 +12,6 @@ using Microsoft.Extensions.Logging;
 
 namespace MVCStore.Areas.Identity.Pages.Account
 {
-    [AllowAnonymous]
     public class LogoutModel : PageModel
     {
         private readonly SignInManager<IdentityUser> _signInManager;
@@ -20,10 +21,6 @@ namespace MVCStore.Areas.Identity.Pages.Account
         {
             _signInManager = signInManager;
             _logger = logger;
-        }
-
-        public void OnGet()
-        {
         }
 
         public async Task<IActionResult> OnPost(string returnUrl = null)
@@ -36,6 +33,8 @@ namespace MVCStore.Areas.Identity.Pages.Account
             }
             else
             {
+                // This needs to be a redirect so that the browser performs a new
+                // request and the identity for the user gets updated.
                 return RedirectToPage();
             }
         }
