@@ -1,20 +1,20 @@
-﻿using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
 using MVCStore.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace MVCStore.Data
 {
-    public class SeedData
+    public static class SeedData
     {
         public static void EnsurePopulated(IApplicationBuilder app)
         {
-            ApplicationDbContext context = app.ApplicationServices
-            .CreateScope().ServiceProvider.GetRequiredService<ApplicationDbContext>();
+            ApplicationDbContext context = app.ApplicationServices.CreateScope().ServiceProvider.GetRequiredService<ApplicationDbContext>();
+
+            if (context.Database.GetPendingMigrations().Any())
+            {
+                context.Database.Migrate();
+            }
+
             if (!context.Products.Any())
             {
                 context.Products.AddRange(
@@ -22,25 +22,24 @@ namespace MVCStore.Data
                 {
                     Name = "Kayak",
                     Description = "A boat for one person",
-                    Category = "Watersports",
                     Price = 275
                 },
                 new Product
                 {
                     Name = "Lifejacket",
                     Description = "Protective and fashionable",
-                    Category = "Watersports",
                     Price = 48.95m
                 },
                 new Product
                 {
                     Name = "Soccer Ball",
                     Description = "FIFA-approved size and weight",
-                    Category = "Soccer",
                     Price = 19.50m
-                });
+                }
+                );
+
+                context.SaveChanges();
             }
-            context.SaveChanges();
         }
     }
 }
