@@ -1,10 +1,4 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Identity;
 
 namespace MVCStore.Data
 {
@@ -12,7 +6,7 @@ namespace MVCStore.Data
     {
         private const string adminEmail = "admin@test.com";
         private const string adminPassword = "Secret123$";
-        public static async Task EnsurePopulated(IApplicationBuilder app)
+        public static async Task EnsurePopulatedAsync(IApplicationBuilder app)
         {
             var serviceProvider = app.ApplicationServices
             .CreateScope().ServiceProvider;
@@ -26,6 +20,24 @@ namespace MVCStore.Data
                 {
                     user = new IdentityUser { UserName = adminEmail, Email = adminEmail };
                     await userManager.CreateAsync(user, adminPassword);
+                }
+
+                var roleManager = serviceProvider.GetRequiredService<RoleManager<IdentityRole>>();
+
+                var roleName = "ProductManagement";
+
+                if (!await roleManager.RoleExistsAsync(roleName))
+                    await roleManager.CreateAsync(new IdentityRole(roleName));
+
+                var adminWithRoleEmail = "adminRole@test.com";
+                var adminWithRolePassword = "Secret123$";
+
+                IdentityUser adminWithRole = await userManager.FindByEmailAsync(adminWithRoleEmail);
+                if (adminWithRole == null)
+                {
+                    adminWithRole = new IdentityUser { UserName = adminWithRoleEmail, Email = adminWithRoleEmail };
+                    await userManager.CreateAsync(adminWithRole, adminWithRolePassword);
+                    await userManager.AddToRoleAsync(adminWithRole, roleName);
                 }
             }
         }
