@@ -1,43 +1,45 @@
-﻿@model IEnumerable<Product>
+﻿@model IEnumerable<MVCStore.Models.Product>
 
 @{
-    ViewBag.Title = "All Products";
-    Layout = "_AdminLayout";
+    ViewBag.Title = "Admin";
+    Layout = "~/Views/Shared/_AdminLayout.cshtml";
 }
 
 <h1>Products</h1>
 
-<a asp-action="Create" class="btn btn-primary mb-3
-   @if(!User.IsInRole("ProductManagement")){
-		@: disabled
-	   }
-   ">Add Product</a>
-
-<table class="table table-striped table-bordered table-sm">
-    <tr>
-        <th class="text-right">ID</th>
-        <th>Name</th>
-        <th class="text-right">Price</th>
-        <th class="text-center">Actions</th>
-    </tr>
-    @foreach (var item in Model)
-    {
+<p>
+    <a asp-action="Create" class="btn btn-primary mb-3">Add Product</a>
+</p>
+ <table class="table table-striped table-bordered table-sm">
         <tr>
-            <td class="text-right">@item.ProductID</td>
-            <td>@item.Name</td>
-            <td class="text-right">@item.Price</td>
-            <td class="text-center">
-                <a asp-action="Edit" class="btn btn-sm btn-warning"
-                   asp-route-productId="@item.ProductID">
-                    Edit
-                </a>
-                <form asp-action="Delete" method="post" style="display: inline">
-                    <input type="hidden" name="ProductId" value="@item.ProductID" />
-                    <button type="submit" class="btn btn-danger btn-sm">
-                        Delete
-                    </button>
-                </form>
-            </td>
+            <th class="text-right">ID</th>
+            <th>Name</th>
+            <th class="text-right">Price</th>
+            <th class="text-center">Actions</th>
         </tr>
-    }
-</table>
+        @foreach (var item in Model)
+        {
+            <tr>
+                <td class="text-right">@item.ProductID</td>
+                <td>@item.Name</td>
+                <td class="text-right">@item.Price</td>
+                <td class="text-center">
+                    <a asp-action="Edit" class="btn btn-sm btn-warning"
+                    asp-route-productId="@item.ProductID">
+                        Edit
+                    </a>
+                    @if (User.IsInRole("ProductManagement"))
+                    {
+                        <form 
+                            asp-action="Delete" 
+                            method="post" style="display: inline">
+                            <input type="hidden" name="ProductId" value="@item.ProductID" />
+                            <button type="submit" class="btn btn-danger btn-sm">
+                                Delete
+                            </button>
+                        </form>
+                    }
+            </td>
+            </tr>
+        }
+    </table>
